requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
scikit-learn==1.3.0
numpy==1.24.0
matplotlib==3.7.0
fastapi==0.104.0
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pillow==10.0.0